Union file.rs in tests and in src in single file and move it in own crate
