+++ Cargo.toml
@@ -15,8 +15,10 @@ anyhow = { version = "1.0.71" }
 configparser = { version = "3.0.2", features = ["indexmap"] }
 queues = { version = "1.0.2" }
 indexmap = { version = "1.9.3" }
+file = { path = "crates/file", version = "0.1.0" }
 
 [dev-dependencies]
 assert_cmd = { version = "2.0.11" }
 predicates = { version = "3.0.3" }
 assert_fs = { version = "1.0.13" }
+file = { path = "crates/file" }
